fix (?) for update_commit_hash.bat

+++ src/L3-solution/BoSSS.Solution/update_commit_hash.bat
@@ -19,16 +19,26 @@ if exist commit_hash.txt (
     set CURRENT_HASH=
 )
 
-echo %CURRENT_HASH%
 
 FOR /F "tokens=* USEBACKQ" %%F IN (`git rev-parse HEAD`) DO (
     SET NEW_HASH=%%F
 )
 
-::echo %NEW_HASH%
+echo new is ---%NEW_HASH%---
+echo cur is ---%CURRENT_HASH%---
 ::echo "equality " "!CURRENT_HASH!" equ "!NEW_HASH!"
 
-if not %CURRENT_HASH% equ %NEW_HASH% (
-    echo "writing new"
+::if not "%CURRENT_HASH%" equ "%NEW_HASH%" (
+::if not "!CURRENT_HASH!" equ "!NEW_HASH!" (
+
+:: be relaxed on trailing spaces
+set "EQUALHASH="
+if "!CURRENT_HASH!" EQU "!NEW_HASH!"  set EQUALHASH=1
+if "!CURRENT_HASH!" EQU "!NEW_HASH! " set EQUALHASH=1
+
+if defined EQUALHASH (
+    echo "existing hash"
+) else (
+	echo "writing new"
     echo !NEW_HASH! > commit_hash.txt
 )
